backend/requirements.txt
blinker==1.6.3
breadability==0.1.20
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
Flask==3.0.0
Flask-Cors==4.0.0
gensim==4.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lxml==4.9.3
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.26.1
pycountry==22.3.5
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
setuptools==68.2.2
smart-open==6.4.0
sumy==0.11.0
tqdm==4.66.1
urllib3==2.0.7
Werkzeug==3.0.0
